used the checkUserCredentialUniqueness during register
