+++ src/main/java/ch/uzh/ifi/hase/soprafs24/security/DriverRegistrationService.java
@@ -81,4 +81,5 @@ public class DriverRegistrationService {
     }
 
     
+    
 }
